# api-docs.json
{"openapi":"3.0.1","info":{"title":"Bank Transaction API","description":"API for managing bank transactions","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/transactions/{id}":{"get":{"tags":["transaction-controller"],"operationId":"getTransactionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}},"put":{"tags":["transaction-controller"],"operationId":"updateTransaction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}},"delete":{"tags":["transaction-controller"],"operationId":"deleteTransaction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/transactions":{"get":{"tags":["transaction-controller"],"operationId":"getAllTransactions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}},"post":{"tags":["transaction-controller"],"operationId":"createTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}}},"/hello":{"get":{"tags":["hello-world-controller"],"operationId":"helloWorld","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Transaction":{"required":["amount","date","description","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"amount":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"},"description":{"maxLength":255,"minLength":0,"type":"string"},"date":{"type":"string","format":"date-time"}}}}}}